finish the monitoring agent and fix some related ddl syntax

+++ linux_sql/sql/ddl.sql
@@ -5,7 +5,7 @@ CREATE TABLE PUBLIC.host_info
      cpu_number       INTEGER NOT NULL,
      cpu_architecture       VARCHAR NOT NULL,
      cpu_model         VARCHAR NOT NULL,
-     cpu_mhz           FLOAT NOT NULL,
+     cpu_mhz           FLOAT(7) NOT NULL,
      L2_cache          INTEGER NOT NULL,
      total_mem         INTEGER NOT NULL,
      "timestamp"       TIMESTAMP NOT NULL
